requirements.txt
opencv-python
Pillow
ultralytics
